Movendo testes do get_backend_info()
